backend/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
wheel==0.45.1
psycopg[binary,pool]==3.2.3
openai==1.59.7
graphene-django==3.2.2